pyproject.toml
[project]
name = "verseminer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "faster-whisper>=1.2.1",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "openai-whisper>=20250625",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.45",
    "torch>=2.9.1",
]
